fix: table configuration delete button problem (#1764)
* fix: table configuration delete button problem

* fix: change copy

* refactor: optimize code

---------

+++ packages/core/client/src/collection-manager/action-hooks.ts
@@ -2,6 +2,7 @@ import { useField, useForm } from '@formily/react';
 import { message } from 'antd';
 import omit from 'lodash/omit';
 import { useEffect } from 'react';
+import { useTranslation } from 'react-i18next';
 import { useCollection, useCollectionManager } from '.';
 import { useCompile } from '..';
 import { useRequest } from '../api-client';
@@ -378,8 +379,12 @@ export const useDestroyAction = () => {
 export const useBulkDestroyAction = () => {
   const { state, setState, refresh } = useResourceActionContext();
   const { resource } = useResourceContext();
+  const { t } = useTranslation();
   return {
     async run() {
+      if (!state?.selectedRowKeys?.length) {
+        return message.error(t('Please select the records you want to delete'));
+      }
       await resource.destroy({
         filterByTk: state?.selectedRowKeys || [],
       });
